backend(Django)/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==3.0
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.7.0
django-crontab==0.7.1
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
idna==3.4
pymongo==3.12.3
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.2.4
urllib3==2.0.2
